services/api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
alembic==1.13.2
psycopg[binary]==3.2.1
python-multipart==0.0.9
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
redis==5.0.8
celery==5.4.0
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
